refactor: 2단계 1차 피드백 반영 - Lottos 로또 구매 금액, 구매 수량 LottoTicket으로 분리
