+++ src/main/java/lotto/Lottos.java
@@ -3,7 +3,6 @@ package lotto;
 import java.util.*;
 
 public class Lottos {
-    private static final int PRICE = 1000;
 
     private List<Lotto> lottos;
 
@@ -35,12 +34,4 @@ public class Lottos {
             lotto.calculateMatchCount(winningLotto);
         }
     }
-
-    public static int countByAmount(int paidAmount) {
-        if (paidAmount < PRICE) {
-            throw new IllegalArgumentException("지불한 금액이 로또 금액보다 적습니다.");
-        }
-        return paidAmount / PRICE;
-    }
-
 }
